Add files via upload
Keyur Padia 22BCE0157 
PROJECT
The driving goal was to make a fully functional code for a pseudo airline management company. The goal was achieved for the most part. Here is a brief description of the aspects we have covered.
Buy/sell planes: This part is the soul of the project. Here the user buys/sells planes out of 15 default choices.
Route management: This part allows you to assign routes amongst 15 default routes to already bought planes. 
Maintenance: This part takes care of the planes already bought. It is divided into 4 types.
Hangar: This shows you the information of all owned planes.
Advertisement: This part enables you to advertise and market for the company.
Catering: This part enables you to organize parties and buffets to campaign for your company.
Stocks: This part is a minigame for the user where he has to buy and sell stocks and try to remain in profit
Logout: This part displays the overall bill for the whole code. Watch out for the easter egg.